requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-test-plus==1.4.0
djangorestframework==3.11.0
flake8==3.8.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient @ file:///C:/Project/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
packaging==20.4
parso==0.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
